+++ app/src/main/java/com/example/tep_timeshareexchangeplatform/UI/Activity/PostingFlow/Adapter/PackagePostingAdapter.kt
@@ -26,7 +26,7 @@ class PackagePostingAdapter: BaseAdapter<PackageModel, PackagePostingAdapter.Pac
 
         }
 
-        fun formatPrice(price: Int): String {
+        fun formatPrice(price: Long): String {
             val formatter = DecimalFormat("#,###")
             return formatter.format(price)
         }
